VehicleMonitoringBE/VehicleMonitoring/Controllers/CustomerController.cs
using Microsoft.AspNetCore.Mvc;
using VehicleMonitoring.Models;

namespace VehicleMonitoring.Controllers
{
    [Route("api/customers")]
    [ApiController]
    public class CustomerController : ControllerBase
    {
        [HttpGet]
        public ActionResult<IEnumerable<Customer>> GetCustomers()
        {
            return DataStore.Customers.ToList();
        }

        [HttpGet("{id}")]
        public ActionResult<Customer> GetCustomer(int id)
        {
            var customer = DataStore.Customers.FirstOrDefault(c => c.Id == id);

            if (customer == null)
            {
                return NotFound();
            }

            return customer;
        }
    }


}
